requirements.txt
# Document Forgery Detection System - No Training Approach
# Core PDF processing
pikepdf>=8.0.0
pdfminer.six>=20221105
pdfplumber>=0.9.0
pdf2image>=1.16.0
PyMuPDF>=1.23.0

# OCR and image processing
pytesseract>=0.3.10
easyocr>=1.7.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
fuzzywuzzy[speedup]>=0.18.0
scikit-image>=0.21.0
scipy>=1.11.0

# AI APIs
openai>=1.0.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=1.0.0

# Data processing and ML utilities
scikit-learn>=1.3.0
xgboost>=1.7.0
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0

# Web frameworks
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
streamlit>=1.28.0
flask>=2.3.0
werkzeug>=2.3.0
plotly>=5.17.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
loguru>=0.7.0

# Enhanced pixel forensics
PyWavelets>=1.4.1

# Optional ML dependencies
# scikit-learn>=1.3.0
# tensorflow>=2.13.0 